tests/wpt/tests/html/semantics/document-metadata/the-link-element/stylesheet-not-removed-until-next-stylesheet-loads.html

<!DOCTYPE html>
<link rel="author" title="Dom Farolino" href="mailto:dom@chromium.org">
<link rel="help" href="https://html.spec.whatwg.org/multipage/links.html#link-type-stylesheet">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<link href="style.css" rel="stylesheet" id="style_test">
<script>
  test(function() {
    assert_true(document.styleSheets.length === 1 &&
                document.styleSheets[0].href.includes("style.css"),
                "The style sheet 'style.css' must be available to scripts");

    style_test.href = "resources/neutral.css?pipe=trickle(d1)";

    assert_true(document.styleSheets.length === 1 &&
                document.styleSheets[0].href.includes("style.css"),
                "The style sheet 'style.css' must remain accessible to " +
                "scripts until its replacement has finished loading");
  }, "Check that a style sheet loaded by a <link> is available until its successor is loaded");
</script>
</head>
</html>
